Security patch
Recent var editing advances allow editing of a refrenced object.
This is generally a good thing.
However, every mob has a client refrence, and admin clients have a refrence, named "holder" to an admin panel object. By editing an admin panel's owner var, the panel would theoretically kick the admin the next time they tried to use it.
Thus, this prevents editing of all admin panel objects, and if you attempt to edit an admin's client, it will announce so and then ask the admin if they would like to allow you to edit their client.