Fixes finished.
Now that all of the old code has been fixed, the next step is to create a system for actually creating the pipes and make it accessable to player in-game. The maploader should work perfectly, but I have returned it to using the built-in one because it is simply faster.